Have the locator also use a proxied port
Previously, the locator was just using the default port of 10334 and map
that for clients to use. Clients could not use 10334, but, instead,
would set up a pool using the mapped, ephemeral port. However, the
locator would still provide its 10334 port to the client and the client
would produce a scary-looking exception since it was unable to connect
to that port.

This commit has the locator using an ephemeral (proxied) port which
ensures that the external (usable by clients) and internal ports are the
same.