Address startup flakiness
Increase the startup timeout to 120 seconds. This is intended to fix the
occasional CI flakiness of containers not starting.

Depend on all servers to start up before signalling that the cluster has
started. This should fix the situation where a region is only being
created on a single server and subsequent GemFire client calls result in
a 'region not found' error.

Allow for verbose test logging by setting the Gradle parameter
'logContainerOutput'. This will log the output of all servers to
stdout.